Fix. chart 코드 수정
